+++ components/Post/PostPings.tsx
@@ -3,6 +3,7 @@ import ProfilePicture from "../ProfilePicture";
 import Link from "next/link";
 import { FormEvent } from "react";
 import PrimaryButton from "../PrimaryButton";
+import { useAuthUser } from "@/contexts/UserContext";
 
 interface PostPingsProps {
   isOpen: boolean;
@@ -10,6 +11,7 @@ interface PostPingsProps {
   isPending: boolean;
   pings: Ping[] | null;
   onToggle: () => void;
+  authorId: string;
 }
 
 export default function PostPings({
@@ -18,8 +20,13 @@ export default function PostPings({
   isPinged,
   pings,
   onToggle,
+  authorId,
 }: PostPingsProps) {
+  const { user } = useAuthUser();
+
   if (!isOpen || pings == null) return null;
+
+  const isOwnPost = authorId === user?.id;
 
   function handleSubmit(e: FormEvent) {
     e.preventDefault();
@@ -30,11 +37,13 @@ export default function PostPings({
     <div className="bg-default-neutral">
       <div className="flex flex-col gap-2 p-2 overflow-auto max-h-60">
         <form onSubmit={handleSubmit} className="flex flex-col">
-          <PrimaryButton disabled={isPending}>
-            {isPinged ? "Unping" : "Ping"} this post
-          </PrimaryButton>
+          {!isOwnPost && (
+            <PrimaryButton disabled={isPending}>
+              {isPinged ? "Unping" : "Ping"} this post
+            </PrimaryButton>
+          )}
         </form>
-
+        {pings.length < 1 && <div>No pings for this post</div>}
         {pings.map((ping) => (
           <Link href={`/profile/~${ping.author.nickname}`} key={ping.createdAt}>
             <div